requirements.txt
# pip install -r requirements.txt
Pillow
Postgres
Django==2.0.2
django-phonenumber-field
psycopg2-binary
django-widget-tweaks
django-import-export
